requirements.txt
flask==3.0.0
google-genai==0.2.0
pyyaml==6.0.1
python-dotenv==1.0.0
flask-swagger-ui==4.11.1
